Conflicts:
	core/baseLanguage/closures/source_gen.caches/jetbrains/mps/baseLanguage/closures/editor/.dependencies
	core/baseLanguage/closures/source_gen/jetbrains/mps/baseLanguage/closures/editor/trace.info
	core/languageDesign/editor/languages/table/source_gen.caches/jetbrains/mps/lang/editor/table/editor/.dependencies
	core/languageDesign/editor/languages/table/source_gen.caches/jetbrains/mps/lang/editor/table/editor/.hash.-12by2gbny8rhxghzalutti8vkdau554
	core/languageDesign/editor/languages/table/source_gen.caches/jetbrains/mps/lang/editor/table/editor/.hash.-3vmbc7cmppsxyl31r5x9lcgmjguinik
	core/languageDesign/editor/languages/table/source_gen.caches/jetbrains/mps/lang/editor/table/editor/.hash.-91ycuxsl24kv1swerlsdnchmzwc5d8b
	core/languageDesign/editor/languages/table/source_gen/jetbrains/mps/lang/editor/table/editor/Transition_Editor.java
	core/languageDesign/editor/languages/table/source_gen/jetbrains/mps/lang/editor/table/editor/trace.info
	core/languageDesign/editor/source_gen.caches/jetbrains/mps/lang/editor/structure/.hash.-7599w7y5xk09vltceoarovjh7tqckm1
	core/languageDesign/editor/source_gen.caches/jetbrains/mps/lang/editor/structure/.hash.-cm0o4r3j7nsqhqjd3t89qztw2lmwx01
	core/languageDesign/editor/source_gen.caches/jetbrains/mps/lang/editor/structure/.hash.df2qa042uick2tf76zj2l2qsc13gk6j
	core/languageDesign/editor/source_gen/jetbrains/mps/lang/editor/editor/trace.info
	core/languageDesign/editor/source_gen/jetbrains/mps/lang/editor/plugin/trace.info

+++ core/languageDesign/editor/source/jetbrains/mps/nodeEditor/cellActions/CellAction_CutNode.java
@@ -15,34 +15,27 @@
  */
 package jetbrains.mps.nodeEditor.cellActions;
 
-import jetbrains.mps.smodel.SNode;
+import jetbrains.mps.baseLanguage.tuples.runtime.Tuples._3;
 import jetbrains.mps.datatransfer.CopyPasteUtil;
-import jetbrains.mps.nodeEditor.EditorCellAction;
 import jetbrains.mps.nodeEditor.EditorContext;
-import jetbrains.mps.nodeEditor.EditorComponent;
-import jetbrains.mps.nodeEditor.NodeRangeSelection;
+import jetbrains.mps.smodel.SNode;
 
 import java.util.List;
-import java.util.LinkedList;
+import java.util.Map;
+import java.util.Set;
 
-public class CellAction_CutNode extends EditorCellAction {
+public class CellAction_CutNode extends CellAction_CopyNode {
 
   public boolean canExecute(EditorContext context) {
-    return context.getNodeEditorComponent().getSelectedCell() != null;
+    return super.canExecute(context);
+    // todo: what about read-only models?
   }
 
   public void execute(EditorContext context) {
-    List<SNode> nodeList = new LinkedList<SNode>();
-    EditorComponent editorComponent = context.getNodeEditorComponent();
-    NodeRangeSelection cellRangeSelection = editorComponent.getNodeRangeSelection();
-    if (cellRangeSelection.isActive()) {
-      nodeList.addAll(cellRangeSelection.getNodes());
-     } else {
-      nodeList.add(editorComponent.getSelectedCell().getSNode());
-    }
-
-    CopyPasteUtil.copyNodesToClipboard(nodeList);
-    for (SNode node : nodeList) {
+    _3<List<SNode>, Map<SNode, Set<SNode>>, String> tuple = extractSelection(context);
+    if (tuple == null) return;
+    CopyPasteUtil.copyNodesAndTextToClipboard(tuple._0(), tuple._1(), tuple._2());
+    for (SNode node : tuple._0()) {
       node.delete();
     }
   }
